cloud_functions/upload_gcs_file/requirements.txt
# empty
google-cloud-storage
pandas
requests
datetime
pytz